requirements.txt
astroid==2.3.3
colorama==0.4.1
Django==2.2.7
django-braces==1.13.0
django-crispy-forms==1.8.0
django-jsonfield==1.3.1
django-model-utils==3.2.0
django-star-ratings==0.9.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
swapper==1.1.1
whitenoise==5.1.0
wrapt==1.11.2
